third_party/blink/web_tests/storage/domstorage/remove-item-expected.txt

Test .removeItem within DOM Storage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing sessionStorage
storage.clear()
PASS storage.length is 0


PASS storage.foo1 is undefined.
storage.foo1 = 'bar'
PASS storage.foo1 is "bar"
storage.removeItem('foo1')
PASS storage.foo1 is undefined.
storage.removeItem('foo1')
PASS storage.foo1 is undefined.


PASS storage['foo2'] is undefined.
storage['foo2'] = 'bar'
PASS storage['foo2'] is "bar"
storage.removeItem('foo2')
PASS storage['foo2'] is undefined.
storage.removeItem('foo2')
PASS storage['foo2'] is undefined.


PASS storage.getItem('foo3') is null
storage.setItem('foo3', 'bar')
PASS storage.getItem('foo3') is "bar"
storage.removeItem('foo3')
PASS storage.getItem('foo3') is null
storage.removeItem('foo3')
PASS storage.getItem('foo3') is null




Testing localStorage
storage.clear()
PASS storage.length is 0


PASS storage.foo1 is undefined.
storage.foo1 = 'bar'
PASS storage.foo1 is "bar"
storage.removeItem('foo1')
PASS storage.foo1 is undefined.
storage.removeItem('foo1')
PASS storage.foo1 is undefined.


PASS storage['foo2'] is undefined.
storage['foo2'] = 'bar'
PASS storage['foo2'] is "bar"
storage.removeItem('foo2')
PASS storage['foo2'] is undefined.
storage.removeItem('foo2')
PASS storage['foo2'] is undefined.


PASS storage.getItem('foo3') is null
storage.setItem('foo3', 'bar')
PASS storage.getItem('foo3') is "bar"
storage.removeItem('foo3')
PASS storage.getItem('foo3') is null
storage.removeItem('foo3')
PASS storage.getItem('foo3') is null
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
